+++ tests/unit/test_solvers.py
@@ -44,6 +44,7 @@ from coreax.kernel import (
 )
 from coreax.solvers import (
     ConditionalKernelHerding,
+    JointRPCholesky,
     KernelHerding,
     MapReduce,
     RandomSample,
@@ -66,6 +67,12 @@ from coreax.util import KeyArrayLike, tree_zero_pad_leading_axis
 
 class _ReduceProblem(NamedTuple):
     dataset: Data
+    solver: Solver
+    expected_coreset: Optional[Coreset] = None
+
+
+class _SupervisedReduceProblem(NamedTuple):
+    dataset: SupervisedData
     solver: Solver
     expected_coreset: Optional[Coreset] = None
 
@@ -430,6 +437,78 @@ class TestRPCholesky(ExplicitSizeSolverTest):
         assert eqx.tree_equal(state, expected_state)
 
 
+class TestJointRPCholesky(ExplicitSizeSolverTest):
+    """Test cases for :class:`coreax.solvers.coresubset.JointRPCholesky`."""
+
+    feature_kernel = SquaredExponentialKernel()
+    response_kernel = PCIMQKernel()
+
+    @override
+    @pytest.fixture(scope="class")
+    def solver_factory(self) -> jtu.Partial:
+        coreset_size = self.shape[0] // 10
+        return jtu.Partial(
+            JointRPCholesky,
+            coreset_size=coreset_size,
+            random_key=self.random_key,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
+        )
+
+    @override
+    @pytest.fixture(params=["random"], scope="class")
+    def reduce_problem(
+        self,
+        request: pytest.FixtureRequest,
+        solver_factory: Union[type[Solver], jtu.Partial],
+    ) -> _SupervisedReduceProblem:
+        if request.param == "random":
+            data_key, supervision_key = jr.split(self.random_key)
+            data = jr.uniform(data_key, self.shape)
+            supervision = jr.uniform(supervision_key, self.shape)
+            solver = solver_factory()
+            expected_coreset = None
+        else:
+            raise ValueError("Invalid fixture parametrization")
+        return _SupervisedReduceProblem(
+            SupervisedData(data, supervision), solver, expected_coreset
+        )
+
+    @override
+    def check_solution_invariants(
+        self, coreset: Coreset, problem: Union[_RefineProblem, _ReduceProblem]
+    ) -> None:
+        """Check functionality of 'unique' in addition to the default checks."""
+        super().check_solution_invariants(coreset, problem)
+        solver = cast(JointRPCholesky, problem.solver)
+        if solver.unique:
+            _, counts = jnp.unique(coreset.nodes.data, return_counts=True)
+            assert max(counts) <= 1
+
+    def test_rpcholesky_state(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Check that the cached JointRPCholesky state is as expected."""
+        dataset, solver, _ = reduce_problem
+        solver = cast(JointRPCholesky, solver)
+        _, state = solver.reduce(dataset)
+
+        x, y = dataset.data, dataset.supervision
+        gramian_diagonal = jax.vmap(solver.kernel.compute_elementwise)((x, y), (x, y))
+        expected_state = RPCholeskyState(gramian_diagonal)
+        assert eqx.tree_equal(state, expected_state)
+
+    def test_non_uniqueness(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
+        dataset, _, _ = reduce_problem
+        solver = JointRPCholesky(
+            random_key=self.random_key,
+            coreset_size=10,
+            feature_kernel=self.feature_kernel,
+            response_kernel=self.response_kernel,
+            unique=False,
+        )
+        solver.reduce(dataset)
+
+
 class TestSteinThinning(RefinementSolverTest, ExplicitSizeSolverTest):
     """Test cases for :class:`coreax.solvers.coresubset.SteinThinning`."""
 
@@ -445,7 +524,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
     """Test cases for :class:`coreax.solvers.coresubset.ConditionalKernelHerding`."""
 
     feature_kernel = SquaredExponentialKernel()
-    response_kernel = SquaredExponentialKernel()
+    response_kernel = PCIMQKernel()
 
     @override
     @pytest.fixture(scope="class")
@@ -465,7 +544,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         self,
         request: pytest.FixtureRequest,
         solver_factory: Union[type[Solver], jtu.Partial],
-    ) -> _ReduceProblem:
+    ) -> _SupervisedReduceProblem:
         if request.param == "random":
             data_key, supervision_key = jr.split(self.random_key)
             data = jr.uniform(data_key, self.shape)
@@ -474,7 +553,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
             expected_coreset = None
         else:
             raise ValueError("Invalid fixture parametrization")
-        return _ReduceProblem(
+        return _SupervisedReduceProblem(
             SupervisedData(data=data, supervision=supervision), solver, expected_coreset
         )
 
@@ -489,12 +568,12 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         scope="class",
     )
     def refine_problem(
-        self, request: pytest.FixtureRequest, reduce_problem: _ReduceProblem
+        self, request: pytest.FixtureRequest, reduce_problem: _SupervisedReduceProblem
     ) -> _RefineProblem:
         """
         Pytest fixture that returns a problem dataset and the expected coreset.
 
-        An expected coreset of 'None' implies the expected coreset for this solver and
+        An expected coreset of `None` implies the expected coreset for this solver and
         dataset combination is unknown.
 
         We expect the '{well,under,over}-sized' cases to return the same result as a
@@ -533,8 +612,8 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         return _RefineProblem(initial_coresubset, solver, expected_coresubset)
 
     # pylint: disable=duplicate-code
-    def test_greedy_cmmd_state(self, reduce_problem: _ReduceProblem) -> None:
-        """Check that the cached ConditionalKernelHerding state is as expected."""
+    def test_greedy_cmmd_state(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Check that the cached `ConditionalKernelHerdingState` is as expected."""
         dataset, solver, _ = reduce_problem
         solver = cast(ConditionalKernelHerding, solver)
         _, state = solver.reduce(dataset)
@@ -572,7 +651,7 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
 
     # pylint: enable=duplicate-code
 
-    def test_batching(self, reduce_problem: _ReduceProblem) -> None:
+    def test_batching(self, reduce_problem: _SupervisedReduceProblem) -> None:
         """Test that batching produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(
@@ -584,8 +663,8 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         )
         solver.reduce(dataset)
 
-    def test_non_uniqueness(self, reduce_problem: _ReduceProblem) -> None:
-        """Test that setting 'unique' to be false produces no errors."""
+    def test_non_uniqueness(self, reduce_problem: _SupervisedReduceProblem) -> None:
+        """Test that setting `unique` to be false produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(
             random_key=self.random_key,
@@ -596,7 +675,9 @@ class TestConditionalKernelHerding(RefinementSolverTest, ExplicitSizeSolverTest)
         )
         solver.reduce(dataset)
 
-    def test_approximate_inverse(self, reduce_problem: _ReduceProblem) -> None:
+    def test_approximate_inverse(
+        self, reduce_problem: _SupervisedReduceProblem
+    ) -> None:
         """Test that using an inverse approximator produces no errors."""
         dataset, _, _ = reduce_problem
         solver = ConditionalKernelHerding(